Restructure project into organized package layout
