+++ buses/base/fipa_message_bus.py
@@ -1,15 +1,18 @@
 from __future__ import annotations
 
-from dataclasses import dataclass, field
-from typing import Dict, List, Optional
-import uuid
+from typing import Dict, TYPE_CHECKING
+
+from messages.fipa_message import FIPAMessage
+
+if TYPE_CHECKING:
+    from agents.base.fipa_agent import FIPAAgent
 
 
 class FipaMessageBus:
     """Simple in-memory message bus for agents."""
 
     def __init__(self) -> None:
-        self.agents: Dict[str, FIPAAgent] = {}
+        self.agents: Dict[str, 'FIPAAgent'] = {}
 
     def register(self, agent: 'FIPAAgent') -> None:
         self.agents[agent.name] = agent
@@ -18,5 +21,3 @@ class FipaMessageBus:
         receiver = self.agents.get(message.receiver)
         if receiver:
             receiver.receive(message)
-
-
